ai-engine/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Security Analysis Tools
slither-analyzer==0.10.0
mythril==0.24.8
py-solc-x==1.1.1

# Machine Learning
scikit-learn==1.3.2
pandas==2.2.2
numpy==1.26.4
joblib==1.3.2
nltk==3.8.1
transformers==4.35.2

# Utilities
requests==2.31.0
aiofiles==24.1.0
psutil==5.9.6
timeout-decorator==0.5.0

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.0.3

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
